SRP - separated CarController & ProductController to separate responsibility of controller
